:truck: chore(api) move get course stuff to `src/evaluation/`
