+++ api/src/evaluation/application/courses/course-controller.js
@@ -1,6 +1,6 @@
+import * as courseSerializer from '../../../shared/infrastructure/serializers/jsonapi/course-serializer.js';
+import { extractUserIdFromRequest } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import * as courseService from '../../domain/services/course-service.js';
-import * as courseSerializer from '../../infrastructure/serializers/jsonapi/course-serializer.js';
-import { extractUserIdFromRequest } from '../../infrastructure/utils/request-response-utils.js';
 
 const get = async function (request, h, dependencies = { courseService, courseSerializer }) {
   const courseId = request.params.id;
