+++ api/src/shared/routes.js
@@ -1,9 +1,8 @@
 import * as assessmentsRoutes from './application/assessments/index.js';
 import * as challengesRoutes from './application/challenges/index.js';
-import * as courses from './application/courses/index.js';
 import * as featureToggles from './application/feature-toggles/index.js';
 import * as healthcheck from './application/healthcheck/index.js';
 
-const sharedRoutes = [healthcheck, assessmentsRoutes, challengesRoutes, courses, featureToggles];
+const sharedRoutes = [healthcheck, assessmentsRoutes, challengesRoutes, featureToggles];
 
 export { sharedRoutes };
