+++ api/tests/evaluation/unit/application/courses/course-controller_test.js
@@ -1,4 +1,4 @@
-import { courseController } from '../../../../../src/shared/application/courses/course-controller.js';
+import { courseController } from '../../../../../src/evaluation/application/courses/course-controller.js';
 import { Course } from '../../../../../src/shared/domain/models/Course.js';
 import { expect, generateAuthenticatedUserRequestHeaders, hFake, sinon } from '../../../../test-helper.js';
