+++ api/tests/evaluation/unit/application/courses/course-route_test.js
@@ -1,6 +1,6 @@
-import { courseController } from '../../../../src/shared/application/courses/course-controller.js';
-import * as moduleUnderTest from '../../../../src/shared/application/courses/index.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import { courseController } from '../../../../../src/evaluation/application/courses/course-controller.js';
+import * as moduleUnderTest from '../../../../../src/evaluation/application/courses/course-route.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Router | course-router', function () {
   describe('GET /api/courses/{id}', function () {
